+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Edge
 * filter_item method for filtering lists in admin views moved here from tog_users
 * Updated template to install preferences plugins (support for users' preferences added to tog_social)
 * Using i18n gem instead of old tog's i18n support
+* Updated template to ask for jquery support (via jrails)
 
 0.6.0
 ----